requirements.txt
# Documents the packages required for WebGnomeAPI

##
## Pyramid Stuff
##
pyramid
waitress

##
## Testing Stuff
##
WebTest
nose
coverage

##
## JSON services stuff
##
git+https://github.com/mozilla-services/cornice.git#egg=cornice
git+https://github.com/Pylons/colander.git#egg=colander

##
## Miscellaneous Stuff
##
awesome-slugify